🔵 Move packages to follow more DDD and hexagonal architecture
