--- src/main/kotlin/com/capitole/capitoleproductcatalogapi/domain/Description.kt
@@ -1,3 +1,0 @@
-package com.capitole.capitoleproductcatalogapi.domain
-
-class Description(val value: String)
