--- src/main/kotlin/com/capitole/capitoleproductcatalogapi/domain/DiscountService.kt
@@ -1,5 +1,0 @@
-package com.capitole.capitoleproductcatalogapi.domain
-
-interface DiscountService {
-  fun getApplicableDiscount(product: Product): DiscountPercentage
-}
