--- src/main/kotlin/com/capitole/capitoleproductcatalogapi/domain/SortField.kt
@@ -1,3 +1,0 @@
-package com.capitole.capitoleproductcatalogapi.domain
-
-enum class SortField { SKU, PRICE, DESCRIPTION, CATEGORY }
