--- src/main/kotlin/com/capitole/capitoleproductcatalogapi/domain/SortOrder.kt
@@ -1,3 +1,0 @@
-package com.capitole.capitoleproductcatalogapi.domain
-
-enum class SortOrder { ASC, DESC }
