+++ src/main/kotlin/com/capitole/capitoleproductcatalogapi/domain/discount/DiscountService.kt
@@ -1,0 +1,7 @@
+package com.capitole.capitoleproductcatalogapi.domain.discount
+
+import com.capitole.capitoleproductcatalogapi.domain.product.Product
+
+interface DiscountService {
+  fun getApplicableDiscount(product: Product): DiscountPercentage
+}
