+++ src/main/kotlin/com/capitole/capitoleproductcatalogapi/domain/product/Description.kt
@@ -1,0 +1,3 @@
+package com.capitole.capitoleproductcatalogapi.domain.product
+
+class Description(val value: String)
