+++ src/main/kotlin/com/capitole/capitoleproductcatalogapi/domain/product/Price.kt
@@ -1,4 +1,4 @@
-package com.capitole.capitoleproductcatalogapi.domain
+package com.capitole.capitoleproductcatalogapi.domain.product
 
 class Price(val value: Double) {
   // TODO price should be positive
