+++ src/main/kotlin/com/capitole/capitoleproductcatalogapi/domain/product/ProductRepository.kt
@@ -1,4 +1,4 @@
-package com.capitole.capitoleproductcatalogapi.domain
+package com.capitole.capitoleproductcatalogapi.domain.product
 
 interface ProductRepository {
     fun findAll(categoryFilter: Category? = null, sortField: SortField?, sortOrder: SortOrder): List<Product>
