+++ src/main/kotlin/com/capitole/capitoleproductcatalogapi/domain/product/SKU.kt
@@ -1,4 +1,4 @@
-package com.capitole.capitoleproductcatalogapi.domain
+package com.capitole.capitoleproductcatalogapi.domain.product
 
 class SKU(val value: String) {
   // TODO SKU should not be blank and should have a valid format
