+++ src/main/kotlin/com/capitole/capitoleproductcatalogapi/domain/product/SortField.kt
@@ -1,0 +1,3 @@
+package com.capitole.capitoleproductcatalogapi.domain.product
+
+enum class SortField { SKU, PRICE, DESCRIPTION, CATEGORY }
