+++ src/main/kotlin/com/capitole/capitoleproductcatalogapi/domain/product/SortOrder.kt
@@ -1,0 +1,3 @@
+package com.capitole.capitoleproductcatalogapi.domain.product
+
+enum class SortOrder { ASC, DESC }
