+++ src/main/kotlin/com/capitole/capitoleproductcatalogapi/infrastructure/configuration/InfrastructureConfiguration.kt
@@ -1,8 +1,8 @@
 package com.capitole.capitoleproductcatalogapi.infrastructure.configuration
 
-import com.capitole.capitoleproductcatalogapi.domain.ProductRepository
-import com.capitole.capitoleproductcatalogapi.infrastructure.repositories.PostgresSQLProductRepository
-import com.capitole.capitoleproductcatalogapi.infrastructure.repositories.ProductSQLBuilder
+import com.capitole.capitoleproductcatalogapi.domain.product.ProductRepository
+import com.capitole.capitoleproductcatalogapi.infrastructure.repositories.product.PostgresSQLProductRepository
+import com.capitole.capitoleproductcatalogapi.infrastructure.repositories.product.ProductSQLBuilder
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate
