+++ src/main/kotlin/com/capitole/capitoleproductcatalogapi/infrastructure/controllers/getproductcatalog/GetProductCatalogResponse.kt
@@ -1,6 +1,6 @@
 package com.capitole.capitoleproductcatalogapi.infrastructure.controllers.getproductcatalog
 
-import com.capitole.capitoleproductcatalogapi.application.ProductCatalogDTO
+import com.capitole.capitoleproductcatalogapi.application.getproductcatalog.ProductCatalogDTO
 
 data class GetProductCatalogResponse(
   val products: List<ProductResponse>
